Split up api.py into multiple views files. Fixed #87.
In order for Flask to recongize all the new routes, we must import the modules at app.py level. I have commented out loading TESTING_PLUGIN for now to avoid circular import.
