+++ minion/backend/app.py
@@ -3,10 +3,17 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 from flask import Flask
-
 app = Flask(__name__)
 
-import minion.backend.api
+import minion.backend.views.base
+import minion.backend.views.groups
+import minion.backend.views.invites
+import minion.backend.views.reports
+import minion.backend.views.users
+import minion.backend.views.scans
+import minion.backend.views.sites
+import minion.backend.views.plans
+import minion.backend.views.plugins
 
 def configure_app(app, production=True, debug=False):
     app.debug = debug
